requirements.txt
streamlit
psutil
pandas
python-dotenv
supabase
rich>=14.1.0